refactor: Consolidate OpenAPI implementation into single file
- Convert existing synthetic-data.ts to OpenAPI format with Zod schemas
- Remove duplicate synthetic-data-openapi.ts file for single source of truth
- Add comprehensive OpenAPI documentation for all endpoints:
  - GET /supported-states - List available state codes
  - POST /generate - Generate synthetic voter data with validation
  - GET /download/:fileId - Download generated files
- Replace manual validation with Zod schema validation
- Maintain full backward compatibility and functionality
- Update index.ts to use consolidated implementation

Benefits:
- Automatic API documentation generation
- Runtime request/response validation
- Type-safe route definitions
- Single maintainable codebase
